refactor(api): extract api methods to file
